- added a quality measure to k-means - fixed bug with sparse centroids - written unit tests for both points above
