+++ src/glib/base/strut.h
@@ -27,7 +27,7 @@ public:
 	static TStr GetStr(const TStrV& StrV, const TStr& DelimiterStr = ",");	
 	static TStr GetStr(const TStrH& StrH, const TStr& FieldDelimiterStr = ":", const TStr& DelimiterStr = ",");	
 	static TStr GetStr(const TFltVV& FltVV, const TStr& DelimiterStr = ",", const TStr& FmtStr = "%g");
-	static TStr GetStr(const TVec<TIntFltKdV>& FltSpVV) { throw TExcept::New("Not implemented!"); }
+	static TStr GetStr(const TVec<TIntFltKdV>& FltSpVV);
 	static TStr GetStr(const TIntIntFltTrV& IntIntFltTrV, const TStr& DelimiterStr = ",", const TStr& FmtStr = "%g");
 	static TStr GetStr(const TUInt64UInt64IntTrV& UInt64UInt64IntTrV, const TStr& DelimiterStr = ",");
 	static TStr GetStr(const TIntSet& IntSet, const TStr& DelimiterStr = ",");
